requirements.txt
streamlit
tensorflow==2.18.0
protobuf>=4.23.0,<6.0.0
gdown
numpy
Pillow
requests
efficientnet # <--- PASTIKAN ADA BARIS INI